+++ twigg-pivotal/lib/twigg-pivotal/pivotal/story.rb
@@ -2,7 +2,8 @@ module Twigg
   module Pivotal
     # Models the story resource in Pivotal Tracker.
     class Story < Resource
-      attr_reader :pivotal_id, :current_state, :story_type, :name, :owned_by
+      attr_reader :pivotal_id, :current_state, :story_type, :name, :url,
+        :owned_by
 
       class << self
         # Returns an array of all open stories for the project identified by
@@ -12,7 +13,7 @@ module Twigg
 
           results = get "projects/#{project_id}/stories",
             filter: 'state:started,finished,delivered,rejected',
-            fields: 'current_state,story_type,name,owned_by'
+            fields: 'current_state,story_type,name,url,owned_by'
 
           results.map { |story| new(story) }
         end
@@ -23,6 +24,7 @@ module Twigg
         raise ArgumentError unless @current_state = json['current_state']
         raise ArgumentError unless @story_type    = json['story_type']
         raise ArgumentError unless @name          = json['name']
+        raise ArgumentError unless @url           = json['url']
 
         # optional (some stories don't have owners)
         @owned_by = json['owned_by']